services/cortex-obs-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cortex-obs-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.110.2",
    "uvicorn[standard]==0.29.0",
    "pydantic==2.7.1",
    "asyncpg==0.29.0",
    "redis==5.0.4",
    "nats-py==2.7.0",
    "hvac==2.1.0",
    "anthropic==0.25.0",
    "httpx==0.27.0",
    "numpy==1.26.4",
    "scikit-learn==1.4.2",
    "opentelemetry-api==1.24.0",
    "opentelemetry-sdk==1.24.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.24.0",
    "prometheus-client==0.20.0",
    "structlog==24.1.0",
]

[tool.uv]
dev-dependencies = ["pytest==8.1.1", "httpx==0.27.0"]
